qe: absorb dml crate into prisma-models
The two crates are no longer distinct in purpose. That simplifies the
dependency graph.
