+++ query-engine/connectors/mongodb-query-connector/src/root_queries/raw.rs
@@ -4,8 +4,7 @@ use mongodb::{
     bson::{from_bson, Bson, Document},
     options::*,
 };
-use prisma_models::dml::PrismaValue;
-use prisma_models::ModelRef;
+use prisma_models::{ModelRef, PrismaValue};
 use std::collections::HashMap;
 
 #[allow(clippy::large_enum_variant)]
