+++ query-engine/connectors/query-connector/src/query_arguments.rs
@@ -1,5 +1,5 @@
 use crate::filter::Filter;
-use prisma_models::*;
+use prisma_models::{ast::FieldArity, *};
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct SkipAndLimit {
@@ -159,7 +159,7 @@ impl QueryArguments {
 
         let has_optional_hop = on_relation.iter().any(|o| {
             o.path.iter().any(|hop| match hop {
-                OrderByHop::Relation(rf) => rf.arity() == dml::FieldArity::Optional,
+                OrderByHop::Relation(rf) => rf.arity() == FieldArity::Optional,
                 OrderByHop::Composite(cf) => !cf.is_required(),
             })
         });
