+++ query-engine/connectors/sql-query-connector/src/database/operations/write.rs
@@ -5,7 +5,7 @@ use crate::{
 };
 use connector_interface::*;
 use itertools::Itertools;
-use prisma_models::{dml::prisma_value::PrismaValue, *};
+use prisma_models::*;
 use quaint::{
     error::ErrorKind,
     prelude::{native_uuid, uuid_to_bin, uuid_to_bin_swapped, Aliasable, Select, SqlFamily},
