--- query-engine/dml/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "dml"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-prisma-value = { path = "../../libs/prisma-value" }
-psl-core = { path = "../../psl/psl-core" }
-
-uuid = { workspace = true, optional = true }
-cuid = { version = "1.2", optional = true }
-nanoid = { version = "0.4.0", optional = true }
-chrono = { version = "0.4.6", features = ["serde"] }
-
-[features]
-# Support for generating default UUID, CUID, nanoid and datetime values. This
-# implies random number generation works, so it won't compile on targets like
-# wasm32.
-default_generators = ["uuid/v4", "cuid", "nanoid"]
