--- query-engine/dml/src/lib.rs
@@ -1,9 +1,0 @@
-//! This code is EOL and nothing new should be added here anymore.
-
-mod default_value;
-mod lift;
-mod native_type_instance;
-
-pub use self::{default_value::*, lift::dml_default_kind, native_type_instance::*};
-pub use prisma_value::{self, PrismaValue};
-pub use psl_core::parser_database::{ast::FieldArity, IndexType, ReferentialAction};
