+++ query-engine/dmmf/src/ast_builders/datamodel_ast_builder.rs
@@ -2,7 +2,7 @@ use crate::serialization_ast::datamodel_ast::{
     Datamodel, Enum, EnumValue, Field, Function, Model, PrimaryKey, UniqueIndex,
 };
 use bigdecimal::ToPrimitive;
-use prisma_models::dml::{self, PrismaValue};
+use prisma_models::{dml_default_kind, encode_bytes, DefaultKind, FieldArity, PrismaValue};
 use psl::{
     parser_database::{walkers, ScalarFieldType},
     schema_ast::ast::WithDocumentation,
@@ -84,14 +84,14 @@ fn composite_type_field_to_dmmf(field: walkers::CompositeTypeFieldWalker<'_>) ->
             ScalarFieldType::Unsupported(_) => unreachable!(),
         },
         db_name: field.mapped_name().map(ToOwned::to_owned),
-        is_required: field.arity() == dml::FieldArity::Required || field.arity() == dml::FieldArity::List,
-        is_list: field.arity() == dml::FieldArity::List,
+        is_required: field.arity() == FieldArity::Required || field.arity() == FieldArity::List,
+        is_list: field.arity() == FieldArity::List,
         is_id: false,
         is_read_only: false,
         has_default_value: field.default_value().is_some(),
         default: field
             .default_value()
-            .map(|dv| default_value_to_serde(&dml::dml_default_kind(dv, field.scalar_type()))),
+            .map(|dv| default_value_to_serde(&dml_default_kind(dv, field.scalar_type()))),
         is_unique: false,
         relation_name: None,
         relation_from_fields: None,
@@ -176,7 +176,7 @@ fn scalar_field_to_dmmf(field: walkers::ScalarFieldWalker<'_>) -> Field {
             ScalarFieldType::Unsupported(_) => unreachable!(),
         },
         is_list: ast_field.arity.is_list(),
-        is_required: matches!(ast_field.arity, dml::FieldArity::Required | dml::FieldArity::List),
+        is_required: matches!(ast_field.arity, FieldArity::Required | FieldArity::List),
         is_unique: !is_id && field.is_unique(),
         is_id,
         is_read_only: field.model().relation_fields().any(|rf| {
@@ -194,7 +194,7 @@ fn scalar_field_to_dmmf(field: walkers::ScalarFieldWalker<'_>) -> Field {
         },
         default: field
             .default_value()
-            .map(|dv| default_value_to_serde(&dml::dml_default_kind(dv.value(), field.scalar_type()))),
+            .map(|dv| default_value_to_serde(&dml_default_kind(dv.value(), field.scalar_type()))),
         relation_name: None,
         relation_from_fields: None,
         relation_to_fields: None,
@@ -212,7 +212,7 @@ fn relation_field_to_dmmf(field: walkers::RelationFieldWalker<'_>) -> Field {
         db_name: None,
         kind: "object",
         is_list: ast_field.arity.is_list(),
-        is_required: matches!(ast_field.arity, dml::FieldArity::Required | dml::FieldArity::List),
+        is_required: matches!(ast_field.arity, FieldArity::Required | FieldArity::List),
         is_unique: false,
         is_id: false,
         is_read_only: false,
@@ -239,10 +239,10 @@ fn relation_field_to_dmmf(field: walkers::RelationFieldWalker<'_>) -> Field {
     }
 }
 
-fn default_value_to_serde(dv: &dml::DefaultKind) -> serde_json::Value {
+fn default_value_to_serde(dv: &DefaultKind) -> serde_json::Value {
     match dv {
-        dml::DefaultKind::Single(value) => prisma_value_to_serde(&value.clone()),
-        dml::DefaultKind::Expression(vg) => {
+        DefaultKind::Single(value) => prisma_value_to_serde(&value.clone()),
+        DefaultKind::Expression(vg) => {
             let args: Vec<_> = vg.args().iter().map(|(_, v)| v.clone()).collect();
             function_to_serde(vg.name(), &args)
         }
@@ -265,7 +265,7 @@ fn prisma_value_to_serde(value: &PrismaValue) -> serde_json::Value {
         PrismaValue::Json(val) => serde_json::Value::String(val.to_string()),
         PrismaValue::Xml(val) => serde_json::Value::String(val.to_string()),
         PrismaValue::List(value_vec) => serde_json::Value::Array(value_vec.iter().map(prisma_value_to_serde).collect()),
-        PrismaValue::Bytes(b) => serde_json::Value::String(dml::prisma_value::encode_bytes(b)),
+        PrismaValue::Bytes(b) => serde_json::Value::String(encode_bytes(b)),
         PrismaValue::Object(pairs) => {
             let mut map = serde_json::Map::with_capacity(pairs.len());
             pairs.iter().for_each(|(key, value)| {
