+++ query-engine/prisma-models/Cargo.toml
@@ -5,11 +5,18 @@ version = "0.0.0"
 
 [dependencies]
 psl.workspace = true
-dml = { path = "../dml" }
 itertools = "0.10"
 prisma-value = { path = "../../libs/prisma-value" }
 bigdecimal = "0.3"
 thiserror = "1.0"
 
+uuid = { workspace = true, optional = true }
+cuid = { version = "1.2", optional = true }
+nanoid = { version = "0.4.0", optional = true }
+chrono = { version = "0.4.6", features = ["serde"] }
+
 [features]
-default_generators = ["dml/default_generators"]
+# Support for generating default UUID, CUID, nanoid and datetime values. This
+# implies random number generation works, so it won't compile on targets like
+# wasm32.
+default_generators = ["uuid/v4", "cuid", "nanoid"]
